Implemented Service layer for Accounts
