+++ Services/IAccountService.cs
@@ -1,0 +1,10 @@
+﻿using HomeBankingAcc.DTOs;
+
+namespace HomeBankingAcc.Services
+{
+    public interface IAccountService
+    {
+        AccountDTO GetAccountById(long id);
+        IEnumerable<AccountDTO> GetAllAccounts();
+    }
+}
