+++ Services/ICardService.cs
@@ -7,9 +7,5 @@ namespace HomeBankingAcc.Services
     {
         CardDTO FindById(long id);
         IEnumerable<CardDTO> GetAllCards();
-        //Task<IEnumerable<Card>> GetCardsByClient(long clientId);
-        //Task<IEnumerable<Card>> FindByType(CardType type, long clientId);
-        //Task<Card> FindByNumber(string number);
-        //Task Save(Card card);
     }
 }
